+++ lib/AppInfo/Application.php
@@ -26,36 +26,60 @@
 
 namespace OCA\CfgShareLinks\AppInfo;
 
+use OCA\CfgShareLinks\Listener\NodeDeletedListener;
+use OCA\CfgShareLinks\Listener\ShareDeletedListener;
 use OCA\CfgShareLinks\Middleware\ShareMiddleware;
+use OCA\Files\Event\LoadAdditionalScriptsEvent;
 use OCP\AppFramework\App;
-use OCP\AppFramework\QueryException;
+use OCP\AppFramework\Bootstrap\IBootContext;
+use OCP\AppFramework\Bootstrap\IBootstrap;
+use OCP\AppFramework\Bootstrap\IRegistrationContext;
 use OCP\EventDispatcher\IEventDispatcher;
+use OCP\Files\Events\Node\NodeDeletedEvent;
+use OCP\Share\Events\ShareDeletedEvent;
 use Psr\Container\ContainerExceptionInterface;
+use Psr\Container\NotFoundExceptionInterface;
 
-class Application extends App {
+class Application extends App implements IBootstrap {
 	public const APP_ID = 'cfg_share_links';
 
-	/**
-	 * @throws QueryException
-	 * @throws ContainerExceptionInterface
-	 */
 	public function __construct() {
 		parent::__construct(self::APP_ID);
+	}
 
-		$container = $this->getContainer();
-
+	public function register(IRegistrationContext $context): void {
 		/**
 		 * Middleware
 		 */
-		$container->registerService('ShareMiddleware', function () {
+		$context->registerService('ShareMiddleware', function () {
 			return new ShareMiddleware();
 		});
-		$container->registerMiddleware('ShareMiddleware');
+		$context->registerMiddleware('ShareMiddleware');
 
-		/* @var IEventDispatcher $dispatcher */
-		$dispatcher = $container->query(IEventDispatcher::class);
-		//        'OCA\Files_Sharing::loadAdditionalScripts'
-		$dispatcher->addListener('OCA\Files::loadAdditionalScripts', function () {
+		/**
+		 * Listen for NodeDeletedEvent (used to remove custom shares of deleted nodes)
+		 */
+		// possible events: BeforeNodeDeletedEvent, MoveToTrashEvent, NodeDeletedEvent
+		$context->registerEventListener(NodeDeletedEvent::class, NodeDeletedListener::class);
+
+		/**
+		 * Listen for ShareDeletedEvent (used to remove custom shares from cfg_shares table)
+		 */
+		$context->registerEventListener(ShareDeletedEvent::class, ShareDeletedListener::class);
+	}
+
+	/**
+	 * @throws ContainerExceptionInterface
+	 * @throws NotFoundExceptionInterface
+	 */
+	public function boot(IBootContext $context): void {
+		/* @var IEventDispatcher $appEventDispatcher */
+		$appEventDispatcher = $context->getAppContainer()->get(IEventDispatcher::class);
+
+		/**
+		 * Load scripts that mount Vue components
+		 */
+		$appEventDispatcher->addListener(LoadAdditionalScriptsEvent::class, function () {
 			script('cfg_share_links', 'cfg_share_links-reg-rename');
 			script('cfg_share_links', 'cfg_share_links-reg-new');
 		});